+++ config/routes.rb
@@ -19,8 +19,6 @@ CompletionKit::Engine.routes.draw do
       post "starters/:key/dismiss", to: "metrics#dismiss_starter", as: :dismiss_starter
     end
     member do
-      post   :add_few_shot
-      delete :remove_few_shot
       post :publish_draft
       post :suggest_variants
       delete :dismiss_suggestion
@@ -89,8 +87,6 @@ CompletionKit::Engine.routes.draw do
         end
         member do
           post :suggest_variants
-          post :add_few_shot
-          delete :remove_few_shot
         end
       end
       resources :metric_groups