+++ db/migrate/20260529000001_remove_few_shot_examples_from_completion_kit_metrics.rb
@@ -1,0 +1,5 @@
+class RemoveFewShotExamplesFromCompletionKitMetrics < ActiveRecord::Migration[8.1]
+  def change
+    remove_column :completion_kit_metrics, :few_shot_examples, :text
+  end
+end