+++ standalone/db/migrate/20260529232016_remove_few_shot_examples_from_completion_kit_metrics.completion_kit.rb
@@ -1,0 +1,6 @@
+# This migration comes from completion_kit (originally 20260529000001)
+class RemoveFewShotExamplesFromCompletionKitMetrics < ActiveRecord::Migration[8.1]
+  def change
+    remove_column :completion_kit_metrics, :few_shot_examples, :text
+  end
+end