+++ standalone/db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[8.1].define(version: 2026_05_28_201817) do
+ActiveRecord::Schema[8.1].define(version: 2026_05_29_232016) do
   create_table "completion_kit_calibrations", force: :cascade do |t|
     t.decimal "corrected_score", precision: 4, scale: 1
     t.datetime "created_at", null: false
@@ -90,7 +90,6 @@ ActiveRecord::Schema[8.1].define(version: 2026_05_28_201817) do
 
   create_table "completion_kit_metrics", force: :cascade do |t|
     t.datetime "created_at", null: false
-    t.text "few_shot_examples"
     t.text "instruction"
     t.string "key"
     t.string "name", null: false